requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --find-links=./wheels --output-file=requirements.txt packages.in
#
--find-links ./wheels

-e libs/datagraph[test]
    # via -r packages.in
-e libs/dux[test]
    # via -r packages.in
-e libs/time[test]
    # via
    #   -r packages.in
    #   aika-datagraph
    #   aika-dux
-e libs/utilities[test]
    # via
    #   -r packages.in
    #   aika-datagraph
    #   aika-time
attrs==21.4.0
    # via
    #   aika-dux
    #   aika-time
    #   pytest
black==22.3.0
    # via aika-dux
click==8.1.3
    # via black
frozendict==2.3.2
    # via aika-dux
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via aika-dux
mongomock==4.0.0
    # via aika-datagraph
mypy-extensions==0.4.3
    # via black
numpy==1.22.3
    # via
    #   aika-datagraph
    #   aika-dux
    #   aika-time
    #   aika-utilities
    #   pandas
packaging==21.3
    # via
    #   mongomock
    #   pytest
pandas==1.4.2
    # via
    #   aika-datagraph
    #   aika-dux
    #   aika-time
    #   aika-utilities
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pymongo==4.1.1
    # via aika-datagraph
pyparsing==3.0.8
    # via packaging
pytest==7.1.2
    # via
    #   aika-datagraph
    #   aika-dux
    #   aika-time
    #   aika-utilities
python-dateutil==2.8.2
    # via pandas
pytz==2022.1
    # via pandas
sentinels==1.0.0
    # via mongomock
six==1.16.0
    # via
    #   mongomock
    #   python-dateutil
tomli==2.0.1
    # via
    #   black
    #   pytest
typing-extensions==4.2.0
    # via black
